requirements.txt
anyio==4.2.0
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-Cors==4.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.1
PyJWT==2.8.0
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-telegram-bot==20.7
pytz==2024.1
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
tzdata==2024.1
Werkzeug==3.0.1
wheel==0.42.0
